requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
asyncpg
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
